% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show map of accidents in a state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number in the}

\item{year}{data year to create a map for}
}
\description{
Plot all accidents for a single year on a map.
}
\examples{
\dontrun{
fars_map_state(35, 2014)
}

}

