% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read raw Fatality Analysis Reporting System File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the full path to the CSV file.}
}
\value{
This function returns a \code{data_frame} object with the same
    column names used in the input data file.
}
\description{
Read a single year of accident data into an R \code{data_frame}. Format a
single years raw data downloaded from US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System into a R object.
The file will be automatically decompressed if it uses the bz2 compression
algorithm.
}
\details{
If the file doesn't exist, an error will be raised.
}
\examples{
\dontrun{
# Read uncompressed file
fars_read('data/accident_2015.csv')
# Read compressed file
fars_read('data/accident_2015.csv.bz2')
}

}

