% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Multiple Years of Accident Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A list of \code{data_frames} containing the month and year of all
   accidents. If a year without data is provided, a warning will show and
   that place in the list will be \code{NULL}.
}
\description{
Read multiple years of Fatality Analysis Reporting System accident data to
produce a list of \code{data_frames} with one accident per record with the
month and year of the accident.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}

}

