% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents across years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A \code{data_frame} with one record per month, and columns for each
        year selected.
}
\description{
Summarize the number of accidents per month across multiple years.
}
\examples{
\dontrun{
fars_summarize_years(c(2013))
fars_summarize_years(c(2013, 2014, 2015))
}

}

