% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate the default filename for a given year of data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The four digit year of data to retrieve.}
}
\value{
The default filename as a string.
}
\description{
Generate the filename used by the Fatality Analysis Reporting System for
the given years worth of data and return as a string.
}
\examples{
make_filename(2013)
make_filename(2014L)
}

