% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function perform the data visualization, for each set of state and years will be displayed the state and
points, where points represents the accidents in this state during the given years.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The operator must know the code of the state to generate the desired map.}

\item{year}{Could be a single year or a list of years.}
}
\value{
The return will be a map with point ploted representing accidents. If the state code is not correct
        an error will be shown saying "invalid STATE number".
}
\description{
This function perform the data visualization, for each set of state and years will be displayed the state and
points, where points represents the accidents in this state during the given years.
}
\examples{
\dontrun{fars_map_state(state.num = 1, year = 2013)}
\dontrun{fars_map_state(state.num = 56, year = 2015)}

}
