% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Week 2 - Assignment}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Is a character string which represents the archive's name.}
}
\value{
This functions returns a data frame of the imported data in case of an invalid name it will return
        a message error "file XXXX does not exist".
}
\description{
This functions will read just only one archive. The input of this function is the filename to be
imported. These archive has data of the US National Highway Traffic Safety Administration's, and
this is a csv file. Keep in mind that the filename will be generate by the make_filename function.
}
\examples{
\dontrun{fars_read(filename = "accident_2013.csv.bz2")}

}
\keyword{coursera}
