% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This functions will import several years by the using of the fars_read function many times. There are some
data manipulation as a creation of a column called year (using mutate) and a selection of two variable per year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Could be a single year or a list of years.}
}
\value{
The results of this function will be a list with only the columns MONTH and year, as you can
        confirm watching the select(MONTH, year). If any year in the list was invalid the function show
        a warning message.
}
\description{
This functions will import several years by the using of the fars_read function many times. There are some
data manipulation as a creation of a column called year (using mutate) and a selection of two variable per year.
}
\examples{
\dontrun{fars_read_years(years = list(2013,2014,2015))}

}
