% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function perform the summarization which consist in the lines number of each dataset, and
bind the separated dataset into one data.frame (using bind_rows). Later transform rows into column
by the spread function. As a result, years in columns rows as months.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Could be a single year or a list pf years.}
}
\value{
The return will be a data.frame with years in columns and months in rows. Each row represent
        the number of accidents.
}
\description{
This function perform the summarization which consist in the lines number of each dataset, and
bind the separated dataset into one data.frame (using bind_rows). Later transform rows into column
by the spread function. As a result, years in columns rows as months.
}
\examples{
\dontrun{fars_summarize_years(years = list(2013,2014,2015))}

}
