% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{This functions will generate the filename string which will be used in the fars_read function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Could be one number or a list of numbers.}
}
\value{
The return of this function could be a single string or a list, depends of its inputs.
}
\description{
This functions will generate the filename string which will be used in the fars_read function
}
\examples{
\dontrun{make_filename(year = 2015)}
\dontrun{make_filename(year = c(2013,2014))}

}
